+++ src/app/dataset-flow/dataset-flow-details/tabs/flow-details-admin-tab/flow-details-admin-tab.component.spec.ts
@@ -1,0 +1,27 @@
+/**
+ * Copyright Kamu Data, Inc. and contributors. All rights reserved.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the LICENSE file.
+ */
+
+import { ComponentFixture, TestBed } from "@angular/core/testing";
+import { FlowDetailsAdminTabComponent } from "./flow-details-admin-tab.component";
+
+describe("FlowDetailsAdminTabComponent", () => {
+    let component: FlowDetailsAdminTabComponent;
+    let fixture: ComponentFixture<FlowDetailsAdminTabComponent>;
+
+    beforeEach(() => {
+        TestBed.configureTestingModule({
+            declarations: [FlowDetailsAdminTabComponent],
+        });
+        fixture = TestBed.createComponent(FlowDetailsAdminTabComponent);
+        component = fixture.componentInstance;
+        fixture.detectChanges();
+    });
+
+    it("should create", () => {
+        expect(component).toBeTruthy();
+    });
+});