+++ src/app/dataset-flow/dataset-flow-details/tabs/flow-details-history-tab/flow-details-history-tab.component.spec.ts
@@ -37,8 +37,10 @@ describe("FlowDetailsHistoryTabComponent", () => {
 
         fixture = TestBed.createComponent(FlowDetailsHistoryTabComponent);
         component = fixture.componentInstance;
-        component.flowHistory = mockFlowHistoryDataFragment;
-        component.flowDetails = mockFlowSummaryDataFragments[2];
+        component.response = {
+            flow: mockFlowSummaryDataFragments[2],
+            flowHistory: mockFlowHistoryDataFragment,
+        };
         fixture.detectChanges();
     });